+++ core-ng-mongo-test/src/test/java/core/framework/mongo/module/TestMongoConfigTest.java
@@ -14,6 +14,7 @@ class TestMongoConfigTest {
     @BeforeEach
     void createTestMongoConfig() {
         config = new TestMongoConfig();
+        config.port = 27017;
     }
 
     @Test