+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## Change log
-### 6.12.4 (4/17/2019 - )
+### 6.12.4 (4/17/2019 - 4/25/2019)
 * mongo: fix: unit test to support multiple test mongo servers on different ports
 * search: fix: elasticsearch local test conflicts with mongo java server
+* json: explicitly defined ZonedDateTime/LocalDateTime format, to make it more strict to comply with ES and js, (put 3 digits for nano fraction)
 
 ### 6.12.3 (4/9/2019 - 4/17/2019)
 * http: update okHTTP to 3.14.1